pyproject.toml
[project]
name = "architekturaheleny-pl"
version = "0.1.0"
description = "architekturaheleny.pl portfolio"
authors = [{ name = "rafsaf", email = "rafal.safin@rafsaf.pl" }]
requires-python = ">=3.14,<3.15"
readme = "README.md"
license = { text = "MIT" }
dependencies = [
    "jinja2 (>=3.1.5)",
    "markdown (>=3.7)",
    "python-dotenv>=1.1.1",
    "watchfiles>=1.1.1",
]

[tool.uv]
package = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
